backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.22
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
plotly==5.22.0
openai==1.30.1
python-dotenv==1.0.1
reportlab==4.2.0
kaleido==0.2.1
httpx==0.27.0
